Update packages
Many classes of the same kind are created.

Classes with similar functionalities are categorised together.

Let's,
* Create new packages to organise the classes
* Refactor the code to accommodate the changes
