+++ src/main/java/duke/command/AddCommand.java
@@ -3,7 +3,7 @@ package duke.command;
 import duke.DukeException;
 import duke.Storage;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 import duke.task.Task;
 
 /**
@@ -24,8 +24,8 @@ public class AddCommand extends Command {
     /**
      * Executes the "add" command, adding the stored task to the task list and updating storage.
      *
-     * @param tasks The task list that the command may operate on.
-     * @param ui The user interface to interact with the user or display messages.
+     * @param tasks   The task list that the command may operate on.
+     * @param ui      The user interface to interact with the user or display messages.
      * @param storage The storage handler to manage data persistence.
      * @throws DukeException If an error occurs while executing the command.
      */
