+++ src/main/java/duke/command/DeleteCommand.java
@@ -3,7 +3,7 @@ package duke.command;
 import duke.DukeException;
 import duke.Storage;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 
 /**
  * Represents a command to delete a task from the task list in the Duke application.
