+++ src/main/java/duke/command/UnmarkCommand.java
@@ -3,7 +3,7 @@ package duke.command;
 import duke.DukeException;
 import duke.Storage;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 
 /**
  * Represents a command to unmark a task.
@@ -23,8 +23,8 @@ public class UnmarkCommand extends Command {
     /**
      * Executes the "unmark" command, and unmarks the task.
      *
-     * @param tasks The task list that the command may operate on.
-     * @param ui The user interface to interact with the user or display messages.
+     * @param tasks   The task list that the command may operate on.
+     * @param ui      The user interface to interact with the user or display messages.
      * @param storage The storage handler to manage data persistence.
      * @throws DukeException If an error occurs while executing the command.
      */
